Move re-exports from `engine` to `prelude`
