+++ src/display.rs
@@ -1,8 +1,7 @@
 //! Enables base64'd output anywhere you might use a `Display` implementation, like a format string.
 //!
 //! ```
-//! use base64::display::Base64Display;
-//! use base64::engine::STANDARD;
+//! use base64::{display::Base64Display, engine::general_purpose::STANDARD};
 //!
 //! let data = vec![0x0, 0x1, 0x2, 0x3];
 //! let wrapper = Base64Display::new(&data, &STANDARD);
@@ -59,7 +58,7 @@ mod tests {
         chunked_encode_matches_normal_encode_random, SinkTestHelper,
     };
     use super::*;
-    use crate::engine::STANDARD;
+    use crate::engine::general_purpose::STANDARD;
 
     #[test]
     fn basic_display() {
